(PC-36934) fix code display on digital bookings
