+++ src/features/bookings/components/Ticket/TicketBottomPart/OnSiteWithdrawal/OnSiteWithdrawal.tsx
@@ -1,8 +1,6 @@
 import React from 'react'
 
-import { TicketCode } from 'features/bookings/components/Ticket/TicketBottomPart/OnSiteWithdrawal/TicketCode'
-import { TicketText } from 'features/bookings/components/Ticket/TicketBottomPart/TicketText'
-import { TicketVisual } from 'features/bookings/components/Ticket/TicketBottomPart/TicketVisual'
+import { TicketCode } from 'features/bookings/components/Ticket/TicketBottomPart/TicketCode'
 import { getDelayMessage } from 'features/bookings/helpers/getDelayMessage'
 
 export const OnSiteWithdrawal = ({
@@ -18,12 +16,5 @@ export const OnSiteWithdrawal = ({
   const displayedDelay = delay ?? ' '
   const text = `Présente le code ci-dessus à l’accueil du lieu indiqué ${displayedDelay}avant le début de l’événement pour récupérer ${isDuo ? 'tes billets' : 'ton billet'}.`
 
-  return (
-    <React.Fragment>
-      <TicketVisual>
-        <TicketCode code={token} />
-      </TicketVisual>
-      <TicketText>{text}</TicketText>
-    </React.Fragment>
-  )
+  return <TicketCode code={token} text={text} />
 }
